Resource monitoring: CPUMonitor is now in-line with spec.

+++ org.osgi.service.resourcemonitoring/src/org/osgi/service/resourcemonitoring/monitor/CPUMonitor.java
@@ -38,6 +38,6 @@ public interface CPUMonitor extends ResourceMonitor {
 	 * 
 	 * @return the CPU usage in nanoseconds
 	 */
-	public int getCPUUsage();
+	public long getCPUUsage();
 
 }
